# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API For Client",
    "description": ""
  },
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization"
    }
  },
  "host": "localhost:4000",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "servers": ["http://localhost:4000", "https://gifts-circle.herokuapp.com"],
...
  "paths": {
...
    "/user/{id}": {
      "get": {
        "description": "",
        "tags": ["User"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/notifications/{id}": {
      "get": {
        "description": "",
        "tags": ["User"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/notifications//{id}": {
      "put": {
        "description": "",
        "tags": ["User"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/changePassword": {
      "post": {
        "description": "",
        "tags": ["User"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/schema/ChangePassword" }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user///{id}": {
      "put": {
        "description": "",
        "tags": ["User"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/schema/UpdateUser" }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/admin/user/{id}": {
      "delete": {
        "description": "",
        "tags": ["Admin_User"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/sourvenir/{id}": {
      "get": {
        "description": "",
        "tags": ["Sourvenir"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/sourvenir/Get/UserSourvenir/{id}": {
      "get": {
        "description": "",
        "tags": ["Sourvenir"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/sourvenir/create": {
      "post": {
        "description": "",
        "tags": ["Sourvenir"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": { "$ref": "#/definitions/schema/Sourvenir" }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/sourvenir/createMany": {
      "post": {
        "description": "",
        "tags": ["Sourvenir"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "array",
            "schema": { "$ref": "#/definitions/schema/Sourvenir" }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/sourvenir/Buy/:id": {
      "put": {
        "description": "",
        "tags": ["Sourvenir"],
        "parameters": [
          {
            "in": "body",
            "required": true,
            "type": "string",
            "name": "id"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/sourvenir/Update": {
      "put": {
        "description": "",
        "tags": ["Sourvenir"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "type": "object",
            "schema": { "$ref": "#/definitions/schema/SourvenirUpdate" }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/sourvenir//{id}": {
      "delete": {
        "description": "",
        "tags": ["Sourvenir"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/marketPlace/Get/All/{id}": {
      "get": {
        "description": "",
        "tags": ["MarketPlace"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/marketPlace/create": {
      "post": {
        "description": "",
        "tags": ["MarketPlace"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "array",
            "schema": { "$ref": "#/definitions/schema/Marketplace" }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
...
}